+++ changelog.md
@@ -5,6 +5,7 @@ Codion Change Log
 ### is.codion.common.core
 - Version.metadata() now returns Optional.
 - Version.build() added, parsing no longer supports space as metadata delimiter.
+- Version, static factory methods replaced with a builder.
 ### is.codion.framework.domain
 - TypeReference added.
 - Attribute.attribute(), Column.column(), EntityType.attribute() and column() overloaded with TypeReference.